requirements.txt
Django==2.0.4
django-map-widgets==0.1.9
geographiclib==1.49
geopy==1.13.0
psycopg2==2.7.4
pytz==2018.4
